Add menu animation on profile vc
