+++ Nexus.xcworkspace/xcuserdata/nickpotts.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -10,7 +10,7 @@
             ignoreCount = "0"
             continueAfterRunningActions = "No"
             filePath = "Nexus/LoginViewController.swift"
-            timestampString = "538711775.179495"
+            timestampString = "538713627.556617"
             startingColumnNumber = "9223372036854775807"
             endingColumnNumber = "9223372036854775807"
             startingLineNumber = "77"
