+++ Nexus/ProfileViewController.swift
@@ -260,32 +260,15 @@ class ProfileViewController: UIViewController {
         handleMenu()
         
     }
+    let menuLaunch = menuLauncher()
     
-    let blackView = UIView()
     func handleMenu() {
         
-        if let window = UIApplication.shared.keyWindow {
-            
-            blackView.backgroundColor = UIColor(white: 0, alpha: 0.5)
-            
-            blackView.addGestureRecognizer(UITapGestureRecognizer(target: self, action: #selector(handleDismissMenu)))
-            
-            window.addSubview(blackView)
-            blackView.frame = window.frame
-            blackView.alpha = 0
-            
-            UIView.animate(withDuration: 0.3, animations: {
-                self.blackView.alpha = 1
-            })
-        }
+        menuLaunch.showMenu()
      
     }
     
-    @objc func handleDismissMenu() {
-        UIView.animate(withDuration: 0.3) {
-            self.blackView.alpha = 0
-        }
-    }
+ 
     /*
     @IBAction func signOut_TouchUpInside(_ sender: Any) {
         print(Auth.auth().currentUser!)
